+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - #506 - ssl_verify_hostname sync
 - #483 - Upgrade dry-validation before releasing 1.3
 - #492 - Use Zeitwerk for code reload in development
+- #508 - Reset the consumers instances upon reconnecting to a cluster
 
 ## 1.2.11
 - [#470](https://github.com/karafka/karafka/issues/470) Karafka not working with dry-configurable 0.8